Update README.md
This is an analysis that has to do with the sales record in different countries as seen from the data attached.  The countries involved Ghana, Cameroon, Nigeria, Niger, Mali, Burkina Faso, CDI and DRC. The grand total of sales is 752, which is distributed or which the count is seen in the different countries involved. From the analysis, it is seen that Ghana makes the highest sales with 197 sales followed by Cameroon with 179 sales.

The products been sold consist of beer, cosmetics and spirits. This sales record was made with analysis to the different type of customers involved, which include Regional Customers, Master Acct, Local Sellers  and Online sellers.